refactor fixed clippy suggestions

+++ src/graphs/regular_grid_3d.rs
@@ -73,14 +73,12 @@ pub fn create_cuboid<F: Clone>(
         for y in 0..size.y {
             for x in 0..size.x {
                 let mut current_neighbours = Vec::new();
-                let pos = ivec3(x as i32, y as i32, z as i32);
+                let pos = ivec3(x, y, z);
                 positions.push(pos + min);
                 for (arc_type, delta) in DIRECTIONS.into_iter().enumerate() {
                     let n_pos = pos + delta;
                     if n_pos.cmpge(IVec3::ZERO).all()
-                        && n_pos
-                            .cmplt(ivec3(size.x as i32, size.y as i32, size.z as i32))
-                            .all()
+                        && n_pos.cmplt(ivec3(size.x, size.y, size.z)).all()
                     {
                         let (i, j, k) = (n_pos.x as usize, n_pos.y as usize, n_pos.z as usize);
                         let index = i + j * size.x as usize + k * size.x as usize * size.y as usize;
@@ -93,7 +91,7 @@ pub fn create_cuboid<F: Clone>(
             }
         }
     }
-    let tiles = iproduct!(0..size.z as i32, 0..size.y as i32, 0..size.x as i32)
+    let tiles = iproduct!(0..size.z, 0..size.y, 0..size.x)
         .enumerate()
         .map(|(i, (z, y, x))| create_node((i, ivec3(x, y, z))))
         .collect_vec();
